[MOD-2505] Android: Fixed "ResultSet" proxy to read integer columns as 64-bit values instead of 32-bit to avoid overflow/underflow
- Improved open() method error handling. (Same change made to Ti.Database.open() in Titanium 8.1.0.)

+++ android/src/appcelerator/encrypteddatabase/TiResultSetProxy.java
@@ -101,7 +101,7 @@ public class TiResultSetProxy extends KrollProxy {
 				if (columnType == android.database.Cursor.FIELD_TYPE_FLOAT) {
 					result = cursor.getDouble(index);
 				} else if (columnType == android.database.Cursor.FIELD_TYPE_INTEGER) {
-					result = cursor.getInt(index);
+					result = cursor.getLong(index);
 				} else if (columnType == android.database.Cursor.FIELD_TYPE_NULL) {
 					result = null;
 				} else if (columnType == android.database.Cursor.FIELD_TYPE_BLOB) {
